requirements.txt
ollama
langchain
langchain-core
langchain_community 
tiktoken 
langchain-nomic
langchain-ollama 
scikit-learn 
langgraph 
tavily-python 
bs4
pytest
pytest-mock
pandas
openpyxl
torch
polars
utils
transformers
datasets
python-dotenv
nltk
sentence-transformers
pymongo
codebleu
tree-sitter-python
astpretty
langchain-community 
langchain-ollama
datasketch
tree-sitter 
tree-sitter-python
langgraph
pyppeteer
graphviz
pygraphviz
textwrap